Processing non-long clicks on application icon.
This CL alone doens't fix the whole bug, see the bug
for the AI list.

In normal mode, clicking at application icon on a task view
is ignored to avoid accidental clicks. To open app properties,
we need a long click.

In accessibility mode, we want to make it easier to open app
properties, and accidental clicks are less likely. Hence, allowing
"simple" clicks.