Various Fixes for Transition
Fix TransitionManager.go not running transition.

Add @TransitionRes for transition resource integers

Reset transitionAlpha when the transition is stopped early.

Bug 19273502

When a transition runs several animators and is interrupted,
some animators may not have been started. Those animators
that haven't been started will not reset their contents
the same as if the animator was canceled. Fade modifies
transitionAlpha and if it is interrupted prior to the
animator starting, the transitionAlpha doesn't get reset.

Make TransitionManager.endTransitions() public.

Make sure Transition constructor is accessible before invoking

Fix Visibility not changing until Animator ends.

Sometimes an animator will be interrupted before it starts
during a transition and the onAnimatorEnd will not be called.
In this case, we must set the final state in the onTransitionEnd.

Fix potential NPE in internal isValidTarget method.

Properly cancel animators when View doesn't exist in starting state.

Bug 19290065

When a transition was interrupted, if a View was not in the start
state, but in the end state, then any animator on that View was
not canceled. This detects that situation and allows canceling
the animator on that View.