Various Transition Fixes
Have the final setVisibility cause a requestLayout.

Suppress Layout during Visibility transitions

Layout while a View is transitioning out can cause strange
behavior during the animation. This CL suppresses the
layout during the transition. It also invalidates
the parent when it has been removed rather than
requesting layout on it.

Fix removeTarget to remove by ID instead of index.

Fix endTransition index out of bounds exception.

When transition.end() is run, it removes itself from
the list of running transitions and perturbs the list.

Fix addTarget not limiting to the given views.

Views matched by instance, or appearing/disappearing, or
named views were not limited.

Add missing addPauseListener.

Fix Fade out during Activity Transition.

The Fade Transition sets the transitionAlpha in onTransitionEnd.
I previously changed Visibility to only do it in onAnimationEnd,
so the onTransitionEnd overrides the Visibility's version. This
fixes that change.

Set the bounds on the drawable for ChangeBounds reparenting.

The bounds for the drawable weren't being set, so the
drawable was never drawn to the correct size.