Various Animation Fixes
Add a check for color type in property values holder

The ValueAnimator's state need to be totally reset during clone.

Reset the Node status when clone.

Being extra careful, reset some more variables on ValueAnimator.

No real bug found related to these set of variables yet, just a clean
up.

Infer value type for keyframes when it is undefined

Fix leak in LayoutTransition

Previously, an OnPreDrawListener was added to clean things up, including
removing the listener, after animations are started.
But if the view is detached before that listener is called, the
listener will remain on that ViewTreeObserver.

This change adds an OnAttachStateChangeListener to perform that cleanup
step when the view is detached, just in case.

Infer value type for ObjectAnimator, too

Previously, changes have been made to infer value types for
PropertyValuesHolder when no value type is defined. This CL applies
the same inferring logic to ObjectAnimator too.

Use ArrayMap instead of HashMap for storing relavant nodes

This fix is to take advantage of ArrayMap, which is a key-value
mapping data structure that is more memory efficient than HashMap.

Clarify docs for Animator.isStarted()

Behavior is different than spec'd for one-shot animators, clarified
docs to reflect that.