Fix bad crop on clip reveal animation
For non-fullscreen apps the dimensions of the app window must be used
to set up the animation.

Fix artifacts in clip reveal animations

clip-reveal the entire screen, not just the app window contents.
Also, account for position of window in non-fullscreen apps.

Make enter reveal animation be zorder top.

This fixes the bug where launcher icons become temporarily visible when
home button is pressed while other activity is starting. By having the
starting activity's animation be zoder top, it will continue to be drawn
on top of the launcher until it animates away.