Not include fullscreen opaque window on screenshot (Combo Fix)
Layout the config-changed window if it has surface
This patch fixes an issue which caused the resumed app to get the
wrong frame at first.

WindowManagerService not propagating X and Y steps
WindowManagerService received X and Y offset steps but failed to pass
them off to the relavent wallpaperservice.

Use the correct parent size to initialize animations
The original logic would use out-of-date parent sizes to initialize
animations. If the screen size is changed after assigning mAnimDw and
mAnimDh in the constructor of WindowStateAnimator, the fromDeltaY
(in the most cases) of TranslateAnimation would be initialized
incorrectly.

In this change, we always use up-to-date parent sizes to initialize
animations to prevent the issue.

Not include fullscreen opaque window on screenshot
Extend the fullscreen meaning by adding opaque drawn check on
screenshot application; if there is a fullscreen window but not
opaque, we will continue finding behind windows; otherwise ending up
including behind windows. It's able to help avoiding capturing BLACK
screenshot bitmap.