Prevent unexpected rotation while going back to keyguard
Refactor to split computeScreenConfigurationLocked

Split into two methods, one that updates the Display and
another that updates the Configuration.

The original logic would let the app hidden by keyguard be able to
decide the orientation. While going back from a show-when-locked app
to keyguard, there would be a short time that keyguard is unable to
decide the orientation, which causes WMS uses the wrong orientation
from the wrong app.