Set non-starting visible resumed activity as resumed activity for the stack
When ensuring visible activities for a stack we check to see if there is
already a resumed activity on the stack, so if we need to start an
activity that needs to be visible, we start it in the paused state
if there is already a resumed activity in the stack. This was't
taking into account a non-starting, visible, and resumed activity.

One case this can happen is if the currently resumed activity in
the stack becomes translucent and the activity behind it needs to
be started because it was previously destroyed due to low memory.

Start non-running activity that should be visible in resume state.

In multi-window mode it is possible for the system-ui launch
multiple activities at the same time that might not all be in a
running process. For the activities that are not in a running
process to be visible, they need to start in the resume state
not the stopped state if they will be the only resumed activity
in the stack.