Volume Motion: Initial show and expand transition.
Also re-enable the content collapse layout transition animations,
supported by ensuring the dialog window is tall enough for long
enough to complete the transition, avoiding clipping.

Volume: Show disabled level on slider when muted by zen.

Volume Dialog: Fix NPE found in monkey testing.

Don't dismiss volume during animation

(sometimes this messes up the state and results in it getting
stuck on the screen)

Remove volume motion delay

Disable window animations, so we can start the show animation directly.

Selectively show zen footer.

Only show zen footer if the active stream is affected by the current
zen mode.

Fix volume expand arrow to be displayed correctly

Sometimes, the volume control expand arrow would be displayed
incorrectly. When different apps use different volume controls and
force different orientations, the position of the arrow (expand button)
will not be updated correctly. When this happens the arrow cannot be
pressed and the volume settings cannot be expanded.

The underlying reason is that onLayoutChange only compares the old
dimensions of a view with the new dimensions, which doesn't take into
account that the last time onLayoutChange was run it may have been run
for a different view (a different volume control), in which case the
dimensions of the new view may not have changed, but the arrow needs to
be repositioned anyway as it needs to be positioned in relation to
another view.

Fix this problem by storing the last stream (volume control) that the
arrow was positioned in relation to, and checking if we're positioning
in relation to the same stream the next time the position of the arrow
is updated.

A11y: Use a backstop dialog timeout when a11y is enabled.

 - Or more precisely, whenever a non-generic-only a11y service
   is enabled.