More state cleanup in View#draw
Simplify state in View#draw(canvas, parent, drawingTime)

Simplify DisplayList / RenderNode state management in
View#draw (the ugly one).

Now, we simply differentiate between drawing with a RenderNode, and
drawing without it. This is because we expect drawing a view while
detached, or to a SW canvas should function the same.

Deprecate and disable legacy caching APIs