Make sure transform matrix is applied when hardware layer is re-created
Don't call updateDisplayListIfDirty outside draw.

Make VPA.setInterpolator(null) unset the interpolator.

Revert "Make VPA.setInterpolator(null) unset the interpolator."

This reverts commit 58ae164100060d1b8a9709aeea7a2b488ff0ac98.

Fix crash during Activity Transition with TextureView.

Bug 20675857

GhostView was synchronously forcing the View's parent to get
the display list when it was created. TextureView's cannot
create a display list until a surface is available and will
crash if it is attempted. This change moves the request to
an invalidate to get it to execute at a later time.

Fix NPE in setSurfaceTexure

Fix whitespace in TextureView

Make VPA.setInterpolator(null) unset the interpolator.

Offset accessibility focus rect for view root's scrollY

Previously we were only partially transforming the focus rect into
window coordinates, so it was offset when the window was panned (for
example, when the IME was showing).

Fix calls to Rect.intersect() in package android.view

This CL checks for the return value for Rect.intersect() for whether
there is actually an intersection before taking the return intersected
rect. In the case of no intersection (Rect.intersect() returns false),
an empty rect will be used as the intersection.

Update doc to include how clipToPadding affects EdgeEffect

Don't add real children for a view with a node provider

Don't display layout bounds of GONE views

Respect accessibility importance for before/after node ordering.

One can set a view to be visited before/after another one for
accessibility but the before/after target views were always added
to the node info for the view regardless if they are included
for accessiilbity. Now the target viwes are reported only if they
should be included for accessibility.

Make WindowInsets#hasInsets check for stable insets as well