Account for IME and screen decors when laying out a resized window.
* Adjust position of target IME resized window if it is been obscured
by the IME.
* Make sure resized window frame is within content frame so it doesn't
extend to the screen decoration regions.

Fixed issue with focused stack frame not displaying in multi-window.

Problem was caused by getting the canvas from the surface while
within a transaction.

Also, cleaned up code a little so it is clearer what is happening.

Don't move multi-window stacks on orientation change.

Change allows resized stacks to remain in the same location
while their content change orientation.

Don't relayout based on a window that isn't visible

The method commitFinishDrawingLocked returned true even if the window
it was called for was hidden. By returning the value that
performShowLocked() returns it only returns true if the window is
shown.

Fixes bug 19100757.

Set TaskStack config orientation based on dimensions

The TaskStack override configuration is set based on the
stack dimensions so we can load the most acturate resources
for activities in the stack in a multi-window environment.

Also, disabled fixed screen orientation for resizeable
activities.

Bug: 19305402

Limit window display frame to size of containing stack

Some widgets like the PopupWindow use the display frame from
WindowManager to figure-out where to place content. We want to
limit the size of the display frame to the containing stack in
a multi-window environment so widgets and apps don't place
contents outside the containing stack of the window.

Make sure window frame fits in the containing frame.

Fixed issue with resized stack sticking to the top of the screen.

WindowState.mUnderStatusBar wasn't set correctly when the stack was
resized before the window was added to it. Changed to use
TaskStack.mUnderStatusBar instead which will always have the
correct answer.