Update JavaDoc of EditorInfo#packageName.
This CL does nothing but updates JavaDoc of EditorInfo#packageName
about the changes made in M.

Report EditorInfo#packageName based on the actual uid.

This is a follow up CL for Ib871141e3381e45c2623c5f4d692da7a7e78fcc5,
which introduced a verification mechanism between the applicaiton uid
and EditorInfo#packageName reported from the application by using
AppOpsManager#checkPackage.

Regression Bug 21408239 caused by above change is really similar to
the situation we saw in Bug 10688644.  As well explained in the CL
description of Ic04f77b3938585b02fccabbc12d2f0dc62b9ef25, the
system has a special ability to run in the processes of other uids.
Hence we need to use Context#getOpPackageName() instead of
Context#getPackageName() in case components declared by the system
is loaded into another process.

Use try-with-resources

We can start relying on try-with-resources in framework.

Call startInput on return from sleep mode

One manifestation of the problem was that input string disappeared when
returning from sleep mode. When editing a TextView with an IME in
landscape orientation, the text would disappear when returning from
sleep mode. The InputMethodManager would be deactivated when the screen
was put into sleep mode as well as the input connection. However when
returning from sleep mode the InputMethodManager was activated, but the
input connection would not be activated again.

The solution is to check focus of the InputMethodManager
which will create a new active input connection to use.

The change is however not specific to this one problem but fundamentally
addresses the issue of lack of startInput on return from sleep mode.