Make UNSPECIFIED measure spec include size hint.
D-pad should be able to push seek bar to 0 or max value

Fix progress tint attributes

Fix for infinite loop in RemoteViewsAdapter

This patch fixes an error in RemoteViewsAdapter when there is only one
view in the cache, and it is bigger than the cache size threshold. This
would cause the cleanup of the cache to get stuck in an infinite loop
while holding the mCache lock that is also needed by for example
getView which is called on the UI thread, leading to ANRs. This patch
breaks the loop when it sees that it can not remove the next view up
for removal.

Consider RTL layout in DPAD navigation in Gallery

Since the Gallery view changes direction the key events LEFT and
RIGHT should swap actions.

Remove unnecessary null check in FastScroller

Add single-arg TextView.setTextAppearance, deprecate two-arg method

The text appearance should always be inflated against the TextView's
original inflation context.

Frameworks/base: Fix AdapterViewAnimator.createOrReuseLayoutParams

The original code was certainly not wrong, but not readable. currentLp
was already an instance (ViewGroup.LayoutParams == LayoutParams), so the
instanceof is the same as a null check. Simplify.

Add a check for invalid spans being set by SpellChecker.

One of the exit conditions of the loop immediately above the code is
when a BreakIterator finishes its input and emits BreakIterator.DONE,
which we were then mistakenly then using as a parameter for setting
spans.