[MULTI] DO NOT MERGE Use correct param for heightMode in RelativeLayout horizontal measure
Update drawable matrix when bounds change in invalidateDrawable()

Avoid NPE when getPositionForView() is called on detached view

Avoid changing internal state of shared drawable

An optimization in ImageView nulled out the internal bitmap of a cached
internal BitmapDrawable object created to wrap a bitmap set on the
ImageView.
However, apps can get ahold of that cached object via
Drawable.getBitmap(),
resulting in having the state of that object they may be using changing
out
from under them.

The change is to null out the cached object when getDrawable() is
called, to avoid
leaking internal state that we may change.
That way, the app can continue to use that object if they want to, but
we are no longer
relying on it internally, and will create a new one when needed.

Issue #22930646  [1P Regression from L] ImageView is blanked out in
Activity in Google Express

Keep OverScroller.SplineOverScroller state consistent.

Currently update() bails out early if it is called
too soon after mStartTime has been set. In this case
mCurrentPosition remains holding an obsolete value
from a previous animation. This causes some strange
behavior (see the referenced bug).

This patch makes sure that mCurrentPosition is
updated every time that mStart is updated.

Revert two changes to LinearLayout measurement

"Include non-zero dimension views in excess space calculation" and
"Always distribute excess space in LinearLayout measurement" changed
LinearLayout behavior significantly in a way that wasn't covered by
CTS tests.

This reverts commits da2f3044092318d44173bdb9918f31c592b86a73 and
4fabc021583cfd5cc551eda7d7b9809683a6ca9a.

Make the RemoteViews context safe for Icons.

Icon needs the RemoteViews' internal context to reflect not
just the Resources of the originating application, but its
package name as well.

Include non-zero dimension views in excess space calculation

Ensure maxHeight is at least 0 during measureChildHorizontal()

Always distribute excess space in LinearLayout measurement

Previously, excess space was added to existing measured dimensions.
This consistently resulted in incorrect allocation of excess space,
since the delta already included the height of any measured children
rather than just the excess space itself.

This CL ensures that excess space is always distributed according to the
layout weights.

Fix handling of RelativeLayout measurement using UNSPECIFIED spec

Always create nodes for views backed by an adapter.

Some views are backed by adapter and if the adapter changed
but views are not updated we were not reporting the views.
This is not correct as the accessibility layer should always
access whatever is on the screen regardless if we know it
will change in the next layout pass.

Improve ImageView drawable re-use

Bug 22403868

Initial attempt only helped if setImageBitmap() was the only
thing called but during new-loading content it's common for a
placeholder to be set via setImageDrawable.

Tweak ImageView slightly to just have a BitmapDrawable that it
lazy-creates but will hold on to for any subsequent calls
to setImageBitmap

docs: Clarified how to find out when a switch is toggled

Per Chris, the existing doc was incorrect: you *can't* check for a
button press and a switch flip the same way. (The button triggers a
click event, but the switch does not.) Chris suggested that we just
remove the reference to onClick and suggest using a listener for
both kinds of buttons (ToggleButton and Switch).

Also pulled one note out of a section where it didn't fit (the bit
about changing a button/switch's state programmatically didn't have
much to do with listening for clicks) and put it at the top, and I
fixed a Javadoc typo for a relevant class that I happened to notice.

See first comment for doc stage location.

Never fully detach header or footer views in AbsListView

AbsListView has special handling of header and footer views that avoids
full attachment; however, we still need to fully detach and reattach
non-header/footer views that cannot be recycled.

Remember focused child during layout when adapter has stable IDs

Ensures behavior is consistent between accessibility focus retention and
keyboard focus retention. Attempts to retain focus when possible.

Allow null child view when checking whether divider should draw

The beginning divider should draw before the first child that is not
set to GONE. Incidentally, there needs to be a check for such a child,
which involves checking the visibility of all the children with lower
indices. This CL takes into account the scenarios where children with
lower indices could be null, in which case we'll treat them as if they
were GONE.

Handle null argument in ImageView.setImageIcon()

Provides consistency with setImageDrawable(). Also updates docs and
annotations for setImageZzz() methods that can take null.

Resolve CompoundButton's foreground drawable for layout direction change

This CL adds a step to resolve CompoundButton's foreground drawable when
layout direction has been changed. Prior this CL, only background
drawable
of the button resolves its layout direciton in respond to system wide
layout direction change. This CL makes the foreground and background
drawables consistent in layout direciton.

Fix logic for whether beginning/end divider should be drawn

Ensure partially-detached AbsListView children are removed

Constrain ListPopupWindow hint width spec to >= 0

Constrain child width and height to >= 0 in FrameLayout.onMeasure()

Constrain width and height to >= 0 in FastScroller layout

Apply thumbTint and trackTint styles to Switch from xml.

Avoid negative childSpecSize in RelativeLayout

Constrain FastScroller max width to be >= 0

Revert "Revert "Work around inconsistent views in
RelativeLayout.onMeasure()""

Turns out this was actually fine.

This reverts commit 9355533bf6f3bcd3fb72fc840253d08ca7dad379.

Ensure child is attached before performing measure/layout in Spinner

Fix javadoc for Adapter.getViewTypeCount()

Remove duplicate words.

Revert "Work around inconsistent views in RelativeLayout.onMeasure()"

This reverts commit 3e2e064a1f989228f48e8faf588393a1a5ee1e31.

Avoid OOBE when AbsListView layout is out of sync with adapter

Views are not synchronized with adapter state until a layout pass
occurs, which may cause an OOBE if a list item is removed and an
accessibility node is obtained before the next layout pass.

This CL caches the item's enabled state on the bound view's layout
params, which allows us to avoid relying on the adapter to populate
accessibility nodes. It also aborts actions if the target position
is no longer valid.

Updates the documentation on AdapterView to reflect that the result
of getPositionForView() may not be synchronized with the adapter.